Determine if Two Trees are Identical (Java)
Given two binary trees, the task is to find if both of them are identical or not. 

Input:
     1          1
   /   \      /   \
  2     3    2     3
Output: Yes
Explanation: There are two trees both
having 3 nodes and 2 edges, both trees
are identical having the root as 1,
left child of 1 is 2 and right child
of 1 is 3.

Expected Time Complexity: O(N).
Expected Auxiliary Space: O(Height of the Tree).