Symmetric Tree (Java)
Given a Binary Tree. Check whether it is Symmetric or not, i.e. whether the binary tree is a Mirror image of itself or not.

Input:
         5
       /   \
      1     1
     /       \
    2         2
Output: True
Explanation: Tree is mirror image of
itself i.e. tree is symmetric

Expected Time Complexity: O(N).
Expected Auxiliary Space: O(Height of the Tree).